CBS: Fix unaligned array allocas
Reuse the same code as in WILoops for padding the row to create
a per-WI aligned inner array.
