+++ tests/workgroup/multilatch_bloop_1_3_1_1.cbs.stdout
@@ -1,0 +1,19 @@
+i:0 0 0 0 before barrier
+i:0 1 0 0 before barrier
+i:0 2 0 0 before barrier
+i:0 0 0 0 after barrier
+i:0 1 0 0 after barrier
+i:0 1 0 0 after latch 1
+i:0 2 0 0 after barrier
+i:0 2 0 0 after latch 1
+i:0 2 0 0 after latch 2
+i:1 0 0 0 before barrier
+i:1 1 0 0 before barrier
+i:1 2 0 0 before barrier
+i:1 0 0 0 after barrier
+i:1 1 0 0 after barrier
+i:1 1 0 0 after latch 1
+i:1 2 0 0 after barrier
+i:1 2 0 0 after latch 1
+i:1 2 0 0 after latch 2
+OK
